moved utils.py to utility/simulation_utils.py
