+++ radp/utility/requirements.txt
@@ -2,3 +2,4 @@ confluent-kafka==2.0.2
 numpy==1.24.2
 pandas==1.5.3
 pyarrow==11.0.0
+torch==1.13.1
